separe roles and tokens
